refactor(server): extract json/html/readBody → http_util.mjs
